app/requirements.txt
fastapi
mangum
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt
pydantic[email]
python-multipart
